Моя попытка № 5

+++ include/task3.h
@@ -1,6 +1,6 @@
 #pragma once
 
-template<typename T, size_t size>
+template<typename T, int size>
 void map(T(&arr)[size],T (*change)(T))
 {
 	for (int i = 0; i < size; i++)
